2.0
It's been a while so I figured I'd revisit this.
It's a shame about the zsh bug where `ztcp -l` socket listeners cannot
be closed (at least as far as I can tell).
On the overall, this should be much more robust.

It's hard to tell how much faster or slower it is though...
On one hand, there's no longer any fork() overhead.
On the other hand, we have to make an extra lookup every call to find
the session's active bucket.

I haven't benchmarked this so far, but I might eventually.